Update Binary_Search_Iterative_Version.cpp
The method used to find the middle element was not correct it may lead to integer overflow in some cases. I implemented the better one.
Here I have quoted what geeksforgeeks says:-
"Some Binary Search implementation looks fine but, the expression “m = (l+r)/2”  fails for large values of l and r. Specifically, it fails if the sum of low and high is greater than the maximum positive int value (2^31 – 1). The sum overflows to a negative value, and the value stays negative when divided by two. In C++ this causes an array index out of bounds with unpredictable results."